requirements.txt
python-telegram-bot==21.7
Flask==3.0.3
requests==2.32.3
